web-platform_22_03_dev/requirements.txt
lxml==4.6.4
Pillow==8.4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0

Django~=4.0
beautifulsoup4~=4.10.0
httplib2~=0.20.2
requests~=2.26.0
openpyxl~=3.0.9
pandas~=1.3.4
pyodbc~=4.0.32
fastkml~=0.12